// star-ui/src/proto/star/v1/firmware_update.ts
export interface FirmwareInfo {
    /**
     * Firmware version string (e.g., "1.0.0").
     *
     * @generated from protobuf field: string version = 1
     */
    version: string;
    /**
     * Git commit hash (short form).
     *
     * @generated from protobuf field: string git_hash = 2
     */
    gitHash: string;
    /**
     * Build date in ISO 8601 format.
     *
     * @generated from protobuf field: string build_date = 3
     */
    buildDate: string;
    /**
     * Firmware size in bytes.
     *
     * @generated from protobuf field: uint32 size = 4
     */
    size: number;
    /**
     * Firmware CRC-32 checksum.
     *
     * @generated from protobuf field: uint32 crc32 = 5
     */
    crc32: number;
    /**
     * IDF version used to build.
     *
     * @generated from protobuf field: string idf_version = 6
     */
    idfVersion: string;
    /**
     * Target chip (e.g., "esp32s3").
     *
     * @generated from protobuf field: string chip_target = 7
     */
    chipTarget: string;
}
// ...
class FirmwareInfo$Type extends MessageType<FirmwareInfo> {
    constructor() {
        super("star.v1.FirmwareInfo", [
            { no: 1, name: "version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "git_hash", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "build_date", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "size", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 5, name: "crc32", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 6, name: "idf_version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 7, name: "chip_target", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<FirmwareInfo>): FirmwareInfo {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.version = "";
        message.gitHash = "";
        message.buildDate = "";
        message.size = 0;
        message.crc32 = 0;
        message.idfVersion = "";
        message.chipTarget = "";
        if (value !== undefined)
            reflectionMergePartial<FirmwareInfo>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FirmwareInfo): FirmwareInfo {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string version */ 1:
                    message.version = reader.string();
                    break;
                case /* string git_hash */ 2:
                    message.gitHash = reader.string();
                    break;
                case /* string build_date */ 3:
                    message.buildDate = reader.string();
                    break;
                case /* uint32 size */ 4:
                    message.size = reader.uint32();
                    break;
                case /* uint32 crc32 */ 5:
                    message.crc32 = reader.uint32();
                    break;
                case /* string idf_version */ 6:
                    message.idfVersion = reader.string();
                    break;
                case /* string chip_target */ 7:
                    message.chipTarget = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FirmwareInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string version = 1; */
        if (message.version !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.version);
        /* string git_hash = 2; */
        if (message.gitHash !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.gitHash);
        /* string build_date = 3; */
        if (message.buildDate !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.buildDate);
        /* uint32 size = 4; */
        if (message.size !== 0)
            writer.tag(4, WireType.Varint).uint32(message.size);
        /* uint32 crc32 = 5; */
        if (message.crc32 !== 0)
            writer.tag(5, WireType.Varint).uint32(message.crc32);
        /* string idf_version = 6; */
        if (message.idfVersion !== "")
            writer.tag(6, WireType.LengthDelimited).string(message.idfVersion);
        /* string chip_target = 7; */
        if (message.chipTarget !== "")
            writer.tag(7, WireType.LengthDelimited).string(message.chipTarget);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message star.v1.FirmwareInfo
 */
export const FirmwareInfo = new FirmwareInfo$Type();